Update satellite.js to consume API using date
Satellite store is initialised with a target date set to the appropriate date for the current point in time. The Timeline component will be responsible for updating the targetDate, and triggering a new store dispatch for satellites. Please also confirm that it is safe to remove the commented out early return in getSatellites - unless I'm missing something, that would prevent any API calls being made after the first one?