Change timeline approach, use performant path sample step, fix path bugs
Creating standalone Cesium Timeline did not allow for manual scrubbing, so instead extract the DOM element and insert into Vue component markup as if we were using jQuery :)

The API calls were getting two sets of orbitals for a single day, so set dates to 0:00, and remove one second from the timescale to ensure we don't get an extra day.

The extra day was messing up the orbital path rendering, which only worked by accident when one orbital so happened to obtain.

Going up to a week (and now a month since I've added that option) was quite heavy when taking 360 samples per day. I've commented out the dynamic step calculation, and plumped for just taking 360 samples for whatever the duration is. So 1 day = most accurate, 1 month (4 weeks) is only approx 13 samples per day, but assuming GEO satellites and with interpolation, this still looks smooth when played back. Since this phase is only concerned with GEO satellites(?) then I figured that this would suffice for now, rather than spend time trying to fine-tune the sample step calculations.