Fix 10 percent drift
The buffer introduced to stop satellites disappearing was causing the positions to drift when viewing more than a day. Also ensured that the iterations of orbital calcs matches the number of days, using the closest orbital element epoch in the available list of orbital elements. May go towards addressing the jerky motion observed when number of available orbital elements is lower than number of days requested.