Handle query string parameters in Cesium component
Enables Cesium component to respond to query string params - filters visible satellites based on comma-separated list in query stirng param, and jumps to timeline point if there is a time set.

TODO:
Date will need watched for and handled elsewhere, and need a way to clear the query string so the user isn't stuck to that time point and those satellites.