Reorganise highlighted entities and selectedEntity tracking
SelectedEntityChanged doesn't fire when the value set doesn't change, but we want to trigger it whenever it's set. So to ensure there is a value change when entity is getting unselected, we set it to either null or false depending on current falsy value.

I also ensure that we don't have highlighted entites and no selection, otherwise the camera remains tracking the last unselected entity, ignoring other highlights.. So when an unselection is happening, and there are still highlighted entities, the camera moves to tracking the last selection.

LastSelectedEntity is there to help us be able to have the highlight tracking / labelling happen in one place rather than spread out over two event handlers. What I maybe missed before was that the selectedEntityChanged listener was firing before the ScreenSpaceEvent, so there is some looping back between those handlers happening now, which makes the whole thing awkward to follow!

It *seems* to work, anyway...