edit: rm glitch when switching satellites
this fixes the glitch, but i'm not sure what functionality we might
lose. its also possible to delay setting trackedEntity with
`setTimeout`, or set it with the flyTo prop complete, but these all
glitch in their own way. entity appears to be a stale object and when
the tracked object is called, there is a brief moment when it tries to
find the old coordinates, before continuing its flyTo path?